style: rename demo component
