+++ docs/vue/components/ComponentA.vue
