+++ docs/vue/components/ComponentB.vue
